Add the client sample paired with openai_api.py
The openai_api.py is just the server sample code of openai. So I write
this client sample to show how-to send the request w/ the api of openai
and get the response from the server. And I also try the benchmark to get
the throughput in words/sencond.

Usage:
1. Start the server by run "python openai_api.py".
2. Start the client with this file with jupyter-notebook.

Signed-off-by: Alex He <heye_dev@163.com>